perbaikan export excel, revisi data management bisa dilihat user

+++ app/Http/Controllers/ManagementRealisasiController.php
@@ -88,6 +88,9 @@ class ManagementRealisasiController extends Controller
 
     public function store(Request $request)
     {
+        if (!(auth()->user()->isSuperAdmin() || auth()->user()->isAdminDirektorat())) {
+            abort(403);
+        }
         $data = $this->validated($request);
 
         $data['direktorat_id'] = $this->resolveDirektoratId($request);
@@ -107,6 +110,9 @@ class ManagementRealisasiController extends Controller
 
     public function update(Request $request, Realisasi $realisasi)
     {
+        if (!(auth()->user()->isSuperAdmin() || auth()->user()->isAdminDirektorat())) {
+            abort(403);
+        }
         $data = $this->validated($request);
 
         $data['direktorat_id'] = $this->resolveDirektoratId($request);
@@ -120,6 +126,9 @@ class ManagementRealisasiController extends Controller
 
     public function destroy(Realisasi $realisasi)
     {
+        if (!(auth()->user()->isSuperAdmin() || auth()->user()->isAdminDirektorat())) {
+            abort(403);
+        }
         $realisasi->delete();
 
         return response()->json([
